Added map button on #second page, and #map page
Added a button on the #second page in the content section. That button
links to a new #map page. I also added a JavaScript file called map.js
in the js folder, and added a script tag for map.js in the head of
index.html. Still working on updating map.js to get the current state's
lat/lon using the global "state" variable.