State Page Changes
Added code on the deviceready event to detect which device is being used
and set the heart image at that point. (The code was on the click event
for the image... which didn't make sense as you need it to run prior to
the click).
Added code to change the play/pause images
Removed the code that started the song right when the page loaded
Changed the replacepage function so that it actually loops through the
JSON data returned. If this was tested with a page with multiple songs
it would have only ever displayed the first song.
Moved the likeButton onclick jquery to inside the replacepage function
so it actually runs.
Added code to remove the click event on the like button after the
initial like occurs.
Random changes to make each of the above work properly.